requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.0.0=pypi_0
aioice=0.9.0=pypi_0
aiortc=1.5.0=pypi_0
altair=5.2.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=23.1.0=pypi_0
av=10.0.0=pypi_0
blinker=1.7.0=pypi_0
brotli=1.1.0=hcfcfb64_1
brotli-bin=1.1.0=hcfcfb64_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2023.11.17=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py311h005e61a_0
cryptography=41.0.7=pypi_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.0=py311h12c1d0e_1
decorator=5.1.1=pyhd8ed1ab_0
dnspython=2.4.2=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fire=0.5.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.45.1=py311ha68e1ae_0
freetype=2.12.1=hdaf720e_2
gast=0.5.4=pypi_0
gettext=0.21.1=h5728263_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glib=2.78.1=h12be248_1
glib-tools=2.78.1=h12be248_1
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.59.3=pypi_0
gst-plugins-base=1.22.7=h001b923_0
gstreamer=1.22.7=hb4038d2_0
h5py=3.10.0=pypi_0
humanfriendly=10.0=pypi_0
icu=73.2=h63175ca_0
idna=3.6=pypi_0
ifaddr=0.2.0=pypi_0
imageio=2.33.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.2.0=h57928b3_50497
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.18.1=pyh5737063_1
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py311h1ea47a8_0
keras=2.15.0=pypi_0
keras2onnx=1.7.0=pypi_0
kiwisolver=1.4.5=py311h005e61a_1
krb5=1.21.2=heb0366b_0
lazy-loader=0.3=pypi_0
lcms2=2.15=h67d730c_3
lerc=4.0.0=h63175ca_0
libblas=3.9.0=20_win64_mkl
libbrotlicommon=1.1.0=hcfcfb64_1
libbrotlidec=1.1.0=hcfcfb64_1
libbrotlienc=1.1.0=hcfcfb64_1
libcblas=3.9.0=20_win64_mkl
libclang=16.0.6=pypi_0
libclang13=15.0.7=default_h77d9078_3
libdeflate=1.19=hcfcfb64_0
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libglib=2.78.1=h16e383f_1
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=h8ffe710_0
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=20_win64_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.44.2=hcfcfb64_0
libtiff=4.6.0=h6e2ebb7_2
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.2=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.11.6=hc3477c8_0
libzlib=1.2.13=hcfcfb64_5
llvmlite=0.41.1=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=py311h1ea47a8_0
matplotlib-base=3.8.2=py311h6e989c2_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=h6a75c08_50497
ml-dtypes=0.2.0=pypi_0
mpmath=1.3.0=pypi_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.26.2=py311h0b4df5a_0
oauthlib=3.2.2=pypi_0
onnx=1.15.0=pypi_0
onnxconverter-common=1.14.0=pypi_0
onnxruntime=1.16.3=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openjpeg=2.5.0=h3d672ee_3
openssl=3.2.0=hcfcfb64_1
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.42=h17e33f8_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=py311h4dd8a23_0
pip=23.3.1=pyhd8ed1ab_0
platformdirs=4.0.0=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.8.0=pypi_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=3.20.2=pypi_0
psutil=5.9.5=py311ha68e1ae_1
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydeck=0.8.1b0=pypi_0
pyee=11.1.0=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pylibsrtp=0.8.0=pypi_0
pymatting=1.1.11=pypi_0
pyopenssl=23.3.0=pypi_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.15.9=py311h125bc19_5
pyqt5-sip=12.12.2=py311h12c1d0e_5
pyreadline3=3.4.1=pypi_0
python=3.11.6=h2628c8c_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytz=2023.3.post1=pypi_0
pywin32=306=py311h12c1d0e_2
pyzmq=25.1.1=py311h9250fbb_2
qt-main=5.15.8=h9e85ed6_17
referencing=0.31.0=pypi_0
rembg=2.0.52=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.22.0=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
sip=6.7.12=py311h12c1d0e_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.28.2=pypi_0
streamlit-webrtc=0.47.1=pypi_0
sympy=1.12=pypi_0
tbb=2021.10.0=h91493d7_2
tenacity=8.2.3=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-intel=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2023.9.26=pypi_0
tk=8.6.13=h5226925_1
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
toolz=0.12.0=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pyhd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=pypi_0
validators=0.22.0=pypi_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
watchdog=3.0.0=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zeromq=4.3.5=h63175ca_0
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.5=h12be248_0
